Added ability to run a global search for a specific phrase
	search [search phrase]

In principle it's no different to `project title foo` other than that the search is across the system rather than confined to a single project.

The search isn't case sensitive, but the ordering of the search terms does matter (as we're just checking for that subphrase within the title).

Can't easily do an issue description search without updating the JILS JSON to include that information (and if I do that, will probably add the issue keywords too)