Added new shortcut functionality, allowing use of issue numbers
When you view an issue or a project listing (including version and component issue listing) the project key will be stored in the cache. You can then simply enter the issue number to print it.

E.g.

	project GPXIN
	10 # Will load GPXIN-10

Also added a set command so that you can manually set the key if needed

	set currproject GPXIN
	10 # Will load GPXIN-10